Refactor uitility method to BigQueryUtils
