+++ src/main/java/io/cdap/delta/bigquery/JobType.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright © 2023 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package io.cdap.delta.bigquery;
+
+/**
+ * Job type which stores additional metadata for the job
+ */
+enum JobType {
+  LOAD_STAGING("stage", false),
+  LOAD_TARGET("load", true),
+  MERGE_TARGET("merge", true);
+
+  private final String id;
+  private final boolean forTargetTable;
+
+  JobType(String id, boolean forTargetTable) {
+    this.id = id;
+    this.forTargetTable = forTargetTable;
+  }
+
+  public String getId() {
+    return id;
+  }
+
+  public boolean isForTargetTable() {
+    return forTargetTable;
+  }
+}
